// src/auth/dto/user-profile.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class UserProfileDto {
  @ApiProperty({ description: '사용자 ID' })
  id: number;

  @ApiProperty({ description: '사용자 이메일' })
  email: string;

  @ApiProperty({ description: '사용자 이름', required: false })
  name?: string;

  @ApiProperty({ description: '휴대폰 번호', required: false })
  phone?: string;

  @ApiProperty({ description: '프로필 조회 성공 메시지' })
  message: string;

  @ApiProperty({ description: '응답 성공 여부' })
  success: boolean;
}